+++ src/pages/common/redirect/components/index.ts
@@ -1,0 +1,1 @@
+export * from './redirect-section';
