+++ src/pages/common/redirect/components/redirect-section/index.ts
@@ -1,0 +1,1 @@
+export { default as RedirectSection } from './RedirectSection';
