+++ src/pages/common/redirect/store/api/index.ts
@@ -1,0 +1,1 @@
+export { getKakaoCallback } from './kakao-callback.api';
