+++ src/pages/common/redirect/store/hooks/index.ts
@@ -1,0 +1,1 @@
+export { useGetKakaoCallback } from './useGetKakaoCallback';
